// src/lib/railway-client/gql/graphql.ts
export enum DeploymentStatus {
  Building = 'BUILDING',
  Crashed = 'CRASHED',
  Deploying = 'DEPLOYING',
  Failed = 'FAILED',
  Initializing = 'INITIALIZING',
  NeedsApproval = 'NEEDS_APPROVAL',
  Queued = 'QUEUED',
  Removed = 'REMOVED',
  Removing = 'REMOVING',
  Skipped = 'SKIPPED',
  Sleeping = 'SLEEPING',
  Success = 'SUCCESS',
  Waiting = 'WAITING'
}